# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Nom de votre API",
        "description": "Description de votre API",
        "contact": {
            "email": "contact@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/categories": {
            "get": {
                "summary": "Get all categories",
                "operationId": "3dc1739ca2bc4527802b0e7c061b36fc",
                "responses": {
                    "200": {
                        "description": "A list of categories"
                    }
                }
            },
            "post": {
                "summary": "Create a new category",
                "operationId": "d48a458b4a6e4fc2e9c50a2e7294e52d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Category 1"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Description of Category 1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The newly created category"
                    }
                }
            }
        },
        "/categories/{id}": {
            "get": {
                "summary": "Get a category by ID",
                "operationId": "5cc175bd1065ea338e65ae4297a394d0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the category to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Details of the specified category"
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            },
            "put": {
                "summary": "Update a category",
                "operationId": "9753f0b68dff9e847003903099280076",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the category to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Category"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated Description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated category"
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete a category by ID",
                "operationId": "bb77efc83d7a5d399d6d8faf9167a471",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the category to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted category"
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            }
        },
...
    },
...
}